Bind DefinedValue data to individual properties
This is needed for e.g. building serialized lists of Serializable classes that have fields marked as DefinedValues. Before, each component or ScriptableObject would override the whole state of dropdowns by whatever the first dropdown in the current Inspector Gui showed for that property, making the of DefinedValues in such a context unfeasible.